Transform class,Gameobject with aggregation principles and has a relation

+++ OverCover/Color.cpp
@@ -1,5 +1,5 @@
 #include "Color.h"
-void Color::SetColor(GLubyte _r,GLubyte _g,GLubyte _b,GLubyte _a)
+void Color::SetColor(GLfloat _r,GLfloat _g,GLfloat _b,GLfloat _a)
 {
 	r=_r;
 	g=_g;
