+++ OverCover/Color.h
@@ -2,7 +2,7 @@
 #include<GL\glew.h>
 struct Color
 {
-	GLubyte r,g,b,a;
-	void SetColor(GLubyte,GLubyte,GLubyte,GLubyte);
+	GLfloat r,g,b,a;
+	void SetColor(GLfloat,GLfloat,GLfloat,GLfloat);
 };
